Getting settings page layout working
Also have been working with how to get and set data from firebase
