+++ src/app/layout/user-nav-bar/user-nav-bar.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { AngularFireAuth } from '@angular/fire/auth';
-import { AngularFirestoreDocument } from '@angular/fire/firestore';
-import { UserProfile } from '@core/models/auth.model';
 import { AuthService } from '@core/services/auth/auth.service';
+import { AngularFirestoreDocument } from '@angular/fire/firestore';
+import { UserProfile } from '@core/models/auth/auth.model';
 import { DataStorageService } from '@core/services/data-storage/dataStorage.service';
 import { Observable } from 'rxjs';
 
@@ -14,31 +14,19 @@ import { Observable } from 'rxjs';
 })
 export class UserNavBarComponent implements OnInit{
 
+  isLogin = false;
   uid: string | undefined;
   constructor(private auth: AuthService, private authfire: AngularFireAuth) {}
 
   ngOnInit() {
     this.authfire.authState.subscribe((x) => {
+      this.isLogin = x != null;
       this.uid = x?.uid
     });
   }
 
-  /*
-  private _document!: AngularFirestoreDocument<UserProfile>;
-  uid: string | undefined;
-  profileInfo!: Observable<UserProfile | undefined>;
-
-  constructor(
-    private dataStorage: DataStorageService,
-    private auth: AngularFireAuth,
-  ) {}
-
-  async ngOnInit() {
-    this.uid = (await this.auth.currentUser)?.uid;
-    if (this.uid !== undefined) {
-      this._document = this.dataStorage.getProfileDetails(this.uid);
-      this.profileInfo = this._document.valueChanges();
-    }
+  signOut() {
+    this.auth.signOut();
   }
-  */
+
 }
