+++ src/app/pages/settings/settings.component.scss
@@ -1,0 +1,4 @@
+.card-bg {
+  background-color: #d3d3d3 ! important;
+  border: #d3d3d3 ! important;
+}
