+++ src/app/pages/user-profile/user-profile.component.html
@@ -16,7 +16,7 @@
 
     <!--Details section w/ all user information-->
     <div class="row">
-      <div class="col-sm-12">
+      <div class="col-sm-12 pt-4">
         <div class="card card-bg shadow-lg">
           <div class="card-body px-4">
