+++ src/app/pages/user-profile/user-profile.component.scss
@@ -1,10 +1,3 @@
-.center {
-  display: block;
-  margin-left: auto;
-  margin-right: auto;
-  width: 50%;
-}
-
 .card-bg {
   background-color: #d3d3d3 ! important;
   border: #d3d3d3 ! important;
